refactor: add db.py and move there a few functions
